% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultility.R
\name{adaptive.stepdown}
\alias{adaptive.stepdown}
\title{the adaptive step-down procedure}
\usage{
adaptive.stepdown(pval, alpha = 0.05)
}
\arguments{
\item{pval}{a vector of p-values (no necessarily sorted)}

\item{alpha}{given FDR level}
}
\value{
a list with num.rejection, the number of rejections computed by this function,
and classifer, a vector of TRUE and FALSE; if TRUE, the corresponding input is regarded
as null, and as non-null if otherwise.
}
\description{
AN ADAPTIVE STEP-DOWN PROCEDURE
}
\details{
This function is a simple implementation of the adaptive step-down procedure
described in Gavrilov et al. (2009)
}
\examples{
adaptive.stepdown(pnorm(-abs(rnorm(1000, c(0, 2)))) * 2)
}
