% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R
\name{bin}
\alias{bin}
\title{Bin the continuous data.}
\usage{
bin(data, order = -5)
}
\arguments{
\item{data}{the observation to be binned.}

\item{order}{see the details}
}
\value{
a list with v be the representative value of each bin and w be the count in the corresponding bin.
}
\description{
Bin the continuous data.
}
\details{
This function bins the continuous data using the equal-width bin in order to
speed up some functions in this package.

h is taken as 10^(-order)

the observations are rounded down to the bins ..., -h, 0, h, ...

To further speed up the process, omit the bin that has 0 count.
}
