% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R
\name{computemixdist}
\alias{computemixdist}
\title{Computing non-parametric mixing distribution}
\usage{
computemixdist(x, ...)
}
\arguments{
\item{x}{a object from implemented family generated by \code{\link{makeobject}}.}

\item{...}{parameters above passed to the specific method}
}
\description{
Computing non-parametric mixing distribution
}
\details{
The full list of implemented family is in \code{\link{makeobject}}.

The avaliable parameters are listed as follows:
\itemize{
\item mix: The initial proper mixing distribution
\item tol: tolerance to stop the code
\item maxiter: maximum iterations allowed.
\item verbose: logical; whether to print the intermediate results.
}

This function essentially calls the class method in the object.
}
\examples{
data = rnorm(500, c(0, 2))
pi0 = 0.5
x = makeobject(data, pi0 = pi0, method = "npnormll")
computemixdist(x)
x = makeobject(data, pi0 = pi0, method = "npnormcvm")
computemixdist(x)
x = makeobject(data, pi0 = pi0, method = "npnormad")
computemixdist(x)
x = makeobject(data, pi0 = pi0, method = "nptll")
computemixdist(x)
}
