% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npnorm.R
\name{ddiscnorm}
\alias{ddiscnorm}
\alias{dnpdiscnorm}
\alias{pdiscnorm}
\alias{pnpdiscnorm}
\title{(non-parametric) discrete normal distribution}
\usage{
ddiscnorm(x, mean = 0, sd = 1, h = 1, log = FALSE)

dnpdiscnorm(
  x,
  mu0 = 0,
  pi0 = 0,
  sd = 1,
  h = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

pdiscnorm(x, mean = 0, sd = 1, h = 1, lower.tail = TRUE, log.p = FALSE)

pnpdiscnorm(
  x,
  mu0 = 0,
  pi0 = 0,
  sd = 1,
  h = 1,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{x}{vector of observations, vector of quantiles}

\item{sd}{standard deviation.}

\item{h}{the discretisation parameter.}

\item{log, log.p}{logical; if TRUE, the result will be given in log scale.}

\item{mu0, mean}{the vector of support points}

\item{pi0}{the vector of weights correponding to the support points \code{mu0}}

\item{lower.tail}{logical; if TRUE, the lower probability is computed}
}
\description{
The density and the distribution function of (non-parametric) discrete normal distribution
}
\details{
\code{ddiscnorm} gives the density, \code{pdiscnorm} gives the distribution function of
the discrete normal distribution. \code{dnpdiscnorm} gives the density, \code{pnpdiscnorm}
gives the distribution function of the non-parametric discrete normal distribution.

The function \code{pnpdiscnorm} uses \code{pnpnorm1} to compute the distribution
function.
}
