% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{extractlower}
\alias{extractlower}
\alias{returnlower}
\title{Extract or returning the lower triangular part of the matrix}
\usage{
extractlower(A)

returnlower(v)
}
\arguments{
\item{A}{a matrix to be extracted the lower triangular part}

\item{v}{a vector to be returned to a symmetric matrix with diagonal 1.}
}
\description{
The function \code{extractlower} is to extract the strict
lower triangular part of a squared matrix and the function
\code{returnlower} is to return the vector value into a
symmetric matrix with diagonal 1.
}
\examples{
a = matrix(1:100, 10, 10)
b = extractlower(a)
d = returnlower(b)
}
