% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultility.R
\name{FDRcontrol}
\alias{FDRcontrol}
\alias{adaptive.stepdown}
\alias{BH}
\title{FDR controlling procedures}
\usage{
adaptive.stepdown(pval, alpha = 0.05)

BH(pval, alpha = 0.05)
}
\arguments{
\item{pval}{a vector of p-values (no necessarily sorted)}

\item{alpha}{given FDR level}
}
\value{
a list with num.rejection, the number of rejections computed by this function,
and classifer, a vector of TRUE and FALSE; if TRUE, the corresponding input is regarded
as null, and as non-null if otherwise.
}
\description{
Implementation of several FDR-controlling procedure in R.
}
\details{
The function \code{adaptive.stepdown} is a simple implementation of the
adaptive step-down procedure described in Gavrilov et al. (2009)

The function \code{BH} is a direct implementation of the procedure
described in Benjamini and Hochberg (1995).
}
\examples{
adaptive.stepdown(pnorm(-abs(rnorm(1000, c(0, 2)))) * 2)
BH(pnorm(-abs(rnorm(1000, c(0, 2)))) * 2)
}
