% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R
\name{lossfunction}
\alias{lossfunction}
\title{loss functions for non-parametric minimisation problems}
\usage{
lossfunction(x, mu0, pi0)
}
\arguments{
\item{x}{a object from implemented family}

\item{mu0}{a vector of support points}

\item{pi0}{a vector of weights corresponding to the support points}
}
\description{
The loss functions for computing loss.
}
\details{
This is a S3 generic function for computing loss function in a minimisation problem.
The full list of implemented family is in \code{\link{makeobject}}.
}
