% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R, R/npnormfcad.R,
%   R/npnormfcadw.R, R/npnormfccvm.R, R/npnormfccvmw.R, R/npnormfcll.R,
%   R/npnormfcllw.R, R/nptfcll.R
\name{makeobject}
\alias{makeobject}
\alias{makeobject.npnormad}
\alias{makeobject.npnormadw}
\alias{makeobject.npnormcvm}
\alias{makeobject.npnormcvmw}
\alias{makeobject.npnormll}
\alias{makeobject.npnormllw}
\alias{makeobject.nptll}
\title{Making object for computation}
\usage{
makeobject(v, mu0 = 0, pi0 = 0, beta = 1, method = "npnormll", ...)

makeobject.npnormad(v, mu0, pi0, beta)

makeobject.npnormadw(v, mu0, pi0, beta, order = -4)

makeobject.npnormcvm(v, mu0, pi0, beta)

makeobject.npnormcvmw(v, mu0, pi0, beta, order = -4)

makeobject.npnormll(v, mu0, pi0, beta)

makeobject.npnormllw(v, mu0, pi0, beta, order = -4)

makeobject.nptll(v, mu0, pi0, beta)
}
\arguments{
\item{v}{the object either numeric or the implmented family}

\item{mu0}{A vector of support points}

\item{pi0}{A vector of weights corresponding to the support points}

\item{beta}{structual parameter.}

\item{method}{An implemented family; see details}

\item{...}{other parameter passed to the constructor.}

\item{order}{the parameter for the binned version.}
}
\description{
These functions are used to make the object for computing the non-paramtric mixing
distribution or estimating the proportion of zero using non-parametric methods.
}
\details{
This is a S3 generic function for making the object for computing the non-parametric
mixing distribution or estimating the proportion of zero.

currently implemented families are:
\itemize{
\item npnormll : normal density using maximum likelihood
\item npnormllw : Binned version of normal density using maximum likelihood
\item npnormcvm : normal density using Cramer-von Mises
\item npnormcvmw : Binned version of normal density using Cramer-von Mises
\item npnormad : normal density using Anderson-Darling
\item npnormadw : Binned version of normal density using Anderson-Darling
\item nptll : t density using maximum likelihood
}

The default method used is npnormll.
}
