% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npnorm.R
\name{dnpnorm}
\alias{dnpnorm}
\alias{pnpnorm}
\alias{pnpnorm1}
\title{non-parametric normal distribution}
\usage{
dnpnorm(x, mu0 = 0, pi0 = 1, sd = 1, log = FALSE)

pnpnorm(x, mu0 = 0, pi0 = 1, sd = 1, lower.tail = TRUE, log.p = FALSE)

pnpnorm1(x, mu0 = 0, pi0 = 1, sd = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{vector of observations, vector of quantiles}

\item{mu0}{the vector of support points}

\item{pi0}{the vector of weights correponding to the support points}

\item{sd}{standard deviation.}

\item{log, log.p}{logical; if TRUE, the result will be given in log scale.}

\item{lower.tail}{logical; if TRUE, the lower probability is computed}
}
\description{
The density and the distribution function of non-parametric normal distribution
}
\details{
\code{dnpnorm} gives the density, \code{pnpnorm} gives the distribution function,
\code{pnpnorm1} focus on the more accurate but slower distribution function of
a non-parametric normal distribution
}
