% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nptll.R
\name{dnpt}
\alias{dnpt}
\alias{pnpt}
\title{non-parametric t distribution}
\usage{
dnpt(x, mu0 = 0, pi0 = 0, df, log = FALSE)

pnpt(x, mu0 = 0, pi0 = 0, df, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{vector of observations, vector of quantiles}

\item{mu0}{the vector of support points}

\item{pi0}{the vector of weights correponding to the support points}

\item{df}{degree of freedom.}

\item{log, log.p}{logical; if TRUE, the result will be given in log scale.}

\item{lower.tail}{logical; if TRUE, the lower probability is computed}
}
\description{
The density and the distribution function of non-parametric t distribution
}
\details{
\code{dnpnorm} gives the density, \code{pnpnorm} gives the distribution function,
}
