% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plotposteriormapping}
\alias{plotposteriormapping}
\title{plot the posterier map}
\usage{
plotposteriormapping(x, result, ...)
}
\arguments{
\item{x}{a vector of observations}

\item{result}{an object of class nspmix}

\item{...}{other parameter passed to \code{plot}}
}
\value{
none
}
\description{
plot the mapping between the original observations and its posterior mean
}
\examples{
n = 100; p = 50
X = matrix(rnorm(n * p), nrow = n, ncol = p)
r = covestEB(X)
x = makeobject(extractlower(r))
r1 = computemixdist(x)
plotposteriormapping(extractlower(r), r1)
}
