% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultility.R
\name{posteriormean}
\alias{posteriormean}
\alias{posteriormean.npnorm}
\alias{posteriormean.npt}
\alias{posteriormean.npnormc}
\title{Find the posterior mean}
\usage{
posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)
}
\arguments{
\item{x}{a vector of observations}

\item{result}{an object of class nspmix}

\item{fun}{the function to transform the mean. It finds the posterior mean
of \code{fun(x)}. The function \code{fun} must be vectorised.}
}
\description{
Find the posterior mean given the observations and the mixing
distribution based on the family in the result
}
\examples{
data = rnorm(500, c(0, 2))
x = makeobject(data, pi0 = 0.5)
r1 = computemixdist(x)
posteriormean(data, r1)
x2 = makeobject(data, pi0 = 0.5, method = "nptll") # equivalent to normal
r2 = computemixdist(x2)
posteriormean(data, r2, fun = function(x) x^2)
data = runif(500, min = -0.5, max = 0.5)
x3 = makeobject(data, method = "npnormcll", beta = 100)
r3 = computemixdist(x3)
posteriormean(data, r3)
}
