% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nptfcll.R
\name{rejectregion.npt}
\alias{rejectregion.npt}
\title{Find the rejection region with t density}
\usage{
rejectregion.npt(result, alpha = 0.05)
}
\arguments{
\item{result}{an object of nptfc}

\item{alpha}{the FDR controlling rate.}
}
\value{
a list with par is the boundary for rejection and area is the propotion of rejection
}
\description{
Find the rejection region with t density
}
\details{
Find the rejection region assuming t component from the nptfc object.
The rejection region is calculated using the density estimate rather than data points hence robust.
The rejection is based on the hypothesis is located at 0.
The optimisation is done via NLopt library (The package nloptr)
}
