% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R, R/npnorm.R, R/nptfcll.R
\name{rejectregion}
\alias{rejectregion}
\alias{rejectregion.npnorm}
\alias{rejectregion.npt}
\title{Find the rejection region}
\usage{
rejectregion(result, alpha = 0.05)

rejectregion.npnorm(result, alpha = 0.05)

rejectregion.npt(result, alpha = 0.05)
}
\arguments{
\item{result}{an object class nspmix}

\item{alpha}{the FDR controlling rate.}
}
\value{
a list with par is the boundary for rejection and area is the propotion of rejection
}
\description{
Find the rejection region
}
\details{
Find the rejection region based on the family in the result
The rejection region is calculated using the density estimate rather than data points hence robust.
The rejection is based on the hypothesis is located at 0.
The optimisation is done via NLopt library (The package nloptr)
}
\examples{
data = rnorm(500, c(0, 2))
x = makeobject(data, pi0 = 0.5)
r1 = computemixdist(x)
rejectregion(r1)
x2 = makeobject(data, pi0 = 0.5, method = "nptll") # equivalent to normal
r2 = computemixdist(x2)
rejectregion(r2)
}
