% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R
\name{computemixdist}
\alias{computemixdist}
\title{Computing non-parametric mixing distribution}
\usage{
computemixdist(x, mix = NULL, tol = 1e-06, maxiter = 100, verbose = FALSE)
}
\arguments{
\item{x}{a object from implemented family generated by \code{\link{makeobject}}.}

\item{mix}{initial mixing distribution (proper)}

\item{tol}{tolerance to stop the code}

\item{maxiter}{maximum iteration to stop the code}

\item{verbose}{logical. If TRUE, the intermediate results will be printed.}
}
\description{
This is a S3 generic function for computing non-parametric mixing distribution
}
\details{
The full list of implemented family is in \code{\link{makeobject}}.
}
\examples{
data = rnorm(500, c(0, 2))
x = makeobject(data, method = "npnormll")
computemixdist(x)
x = makeobject(data, method = "npnormcvm")
computemixdist(x)
x = makeobject(data, method = "npnormad")
computemixdist(x)
x = makeobject(data, method = "nptll")
computemixdist(x)
}
