% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{covestEB}
\alias{covestEB}
\title{Estimating Covariance Matrix using Empirical Bayes}
\usage{
covestEB(X, estpi0 = FALSE, order = -3, verbose = FALSE)
}
\arguments{
\item{X}{a matrix of size n * p, where n is the number of observations and p is the number of variables}

\item{estpi0}{logical; if TRUE, the NPMLE is estimated based on the estimation of pi0, which in this case
can be used to detect sparsity or assume sparsity.}

\item{order}{the level of binning to use when the number of observations passed to the computation is greater
than 5000.}

\item{verbose}{logical; If TRUE, the intermediate results will be shown.}
}
\value{
a covariance matrix estimate of size p * p.
}
\description{
Estimating covariance matrix using Empirical Bayes
}
\examples{
n = 100; p = 50
X = matrix(rnorm(n * p), nrow = n, ncol = p)
r = covestEB(X)
}
