% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R, R/npnormad.R, R/npnormcvm.R,
%   R/npnormll.R, R/nptll.R
\name{estpi0}
\alias{estpi0}
\alias{estpi0.npnormad}
\alias{estpi0.npnormcvm}
\alias{estpi0.npnormll}
\alias{estpi0.nptll}
\title{Computing non-parametric mixing distribution with estimated proportion at 0}
\usage{
estpi0(x, ...)

estpi0.npnormad(
  x,
  val = qAD(0.5, lower.tail = FALSE),
  mix = NULL,
  tol = 1e-06,
  maxiter = 100,
  verbose = FALSE
)

estpi0.npnormcvm(
  x,
  val = qCvM(0.5, lower.tail = FALSE),
  mix = NULL,
  tol = 1e-06,
  maxiter = 100,
  verbose = FALSE
)

estpi0.npnormll(
  x,
  val = -log(0.5),
  mix = NULL,
  tol = 1e-06,
  maxiter = 100,
  verbose = FALSE
)

estpi0.nptll(
  x,
  val = -log(0.5),
  mix = NULL,
  tol = 1e-06,
  maxiter = 100,
  verbose = FALSE
)
}
\arguments{
\item{x}{a object from implemented family}

\item{...}{parameters passed to the specific method.}

\item{val}{the threshold value.}

\item{mix}{initial mixing distribution (proper)}

\item{tol}{tolerance to stop the code}

\item{maxiter}{maximum iteration to stop the code}

\item{verbose}{logical. If TRUE, the intermediate results will be printed.}
}
\description{
computing non-parametric mixing distribution with estimated proportion at 0
}
\details{
This is a S3 generic function for computing non-parametric mixing
distribution with estimated proportion at 0. Different families will
have different threshold values.

The full list of implemented family is in \code{\link{makeobject}}.
}
\examples{
data = rnorm(500, c(0, 2))
pi0 = 0.5
x = makeobject(data, method = "npnormll")
estpi0(x)
x = makeobject(data, method = "npnormcvm")
estpi0(x)
x = makeobject(data, method = "npnormad")
estpi0(x)
}
