% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R, R/npnormfcad.R,
%   R/npnormfcadw.R, R/npnormfccvm.R, R/npnormfccvmw.R, R/npnormfcll.R,
%   R/npnormfcllw.R
\name{estpi0}
\alias{estpi0}
\alias{estpi0.npnormad}
\alias{estpi0.npnormadw}
\alias{estpi0.npnormcvm}
\alias{estpi0.npnormcvmw}
\alias{estpi0.npnormll}
\alias{estpi0.npnormllw}
\title{Computing non-parametric mixing distribution with estimated proportion at 0}
\usage{
estpi0(x, val, mix, tol, maxiter, verbose)

\method{estpi0}{npnormad}(
  x,
  val = qAD(0.05, lower.tail = FALSE),
  mix = NULL,
  tol = 1e-06,
  maxiter = 100,
  verbose = FALSE
)

\method{estpi0}{npnormadw}(
  x,
  val = qAD(0.05, lower.tail = FALSE),
  mix = NULL,
  tol = 1e-06,
  maxiter = 100,
  verbose = FALSE
)

\method{estpi0}{npnormcvm}(
  x,
  val = qCvM(0.05, lower.tail = FALSE),
  mix = NULL,
  tol = 1e-06,
  maxiter = 100,
  verbose = FALSE
)

\method{estpi0}{npnormcvmw}(
  x,
  val = qCvM(0.05, lower.tail = FALSE),
  mix = NULL,
  tol = 1e-06,
  maxiter = 100,
  verbose = FALSE
)

\method{estpi0}{npnormll}(
  x,
  val = 0.5 * log(length(x$v)),
  mix = NULL,
  tol = 1e-06,
  maxiter = 100,
  verbose = FALSE
)

\method{estpi0}{npnormllw}(
  x,
  val = 0.5 * log(sum(x$w)),
  mix = NULL,
  tol = 1e-06,
  maxiter = 100,
  verbose = FALSE
)
}
\arguments{
\item{x}{a object from implemented family}

\item{val}{the threshold value.}

\item{mix}{initial mixing distribution (proper)}

\item{tol}{tolerance to stop the code}

\item{maxiter}{maximum iteration to stop the code}

\item{verbose}{logical. If TRUE, the intermediate results will be printed.}
}
\description{
computing non-parametric mixing distribution with estimated proportion at 0
}
\details{
This is a S3 generic function for computing non-parametric mixing
distribution with estimated proportion at 0. Different families will
have different threshold value.
}
