% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R
\name{gradientfunction}
\alias{gradientfunction}
\title{The gradient function}
\usage{
gradientfunction(x, mu, mu0, pi0, order)
}
\arguments{
\item{x}{a object from implemented family}

\item{mu}{a vector of locations to look for}

\item{mu0}{a vector of support points}

\item{pi0}{a vector of weights corresponding to the support points}

\item{order}{a vector of length 3 with either 1 or 0. 1 to compute, 0 not to compute.}
}
\value{
a list of length 3. d0, d1, d2 correspond to the order.
}
\description{
gradient function
}
\details{
This is a S3 generic function returning gradident function with order specified as a list
}
