% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R
\name{lossfunction}
\alias{lossfunction}
\title{loss function}
\usage{
lossfunction(x, mu0, pi0)
}
\arguments{
\item{x}{a object from implemented family}

\item{mu0}{a vector of support points}

\item{pi0}{a vector of weights corresponding to the support points}
}
\description{
loss function
}
\details{
This is a S3 generic loss function for computing minimisation problem in order to find
non-parametric estimator
}
