% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R
\name{makeobject}
\alias{makeobject}
\title{Making object for computation}
\usage{
makeobject(v, method = "npnormll", ...)
}
\arguments{
\item{v}{the object either numeric or the implmented family}

\item{method}{An implemented family; see details}

\item{...}{other parameter passed to the constructor.}

\item{mu0}{A vector of support points}

\item{pi0}{A vector of weights corresponding to the support points}

\item{beta}{structual parameter.}

\item{order}{the parameter for the binned version.}
}
\description{
These functions are used to make the object for computing the non-paramtric mixing
distribution or estimating the proportion of zero using non-parametric methods.
}
\details{
This is a S3 generic function for making the object for computing the non-parametric
mixing distribution or estimating the proportion of zero.

currently implemented families are:
\itemize{
\item npnormll : normal density using maximum likelihood
\item npnormllw : Binned version of normal density using maximum likelihood
\item npnormcvm : normal density using Cramer-von Mises
\item npnormcvmw : Binned version of normal density using Cramer-von Mises
\item npnormad : normal density using Anderson-Darling
\item npnormadw : Binned version of normal density using Anderson-Darling
\item nptll : t density using maximum likelihood
\item npnormcll : one-parameter normal distribution used for approximate sample
correlation coefficients using maximum likelihood. This does not have a
corresponding estimation of zero due to incomplete of theory.
}

The default method used is npnormll.
}
