% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R, R/npnormad.R, R/npnormadw.R,
%   R/npnormcll.R, R/npnormcvm.R, R/npnormcvmw.R, R/npnormll.R, R/npnormllw.R,
%   R/nptll.R
\name{makeobject}
\alias{makeobject}
\alias{makeobject.npnormad}
\alias{makeobject.npnormadw}
\alias{makeobject.npnormcll}
\alias{makeobject.npnormcvm}
\alias{makeobject.npnormcvmw}
\alias{makeobject.npnormll}
\alias{makeobject.npnormllw}
\alias{makeobject.nptll}
\title{Making object for computation}
\usage{
makeobject(v, method = "npnormll", ...)

makeobject.npnormad(v, mu0, pi0, beta, order)

makeobject.npnormadw(v, mu0, pi0, beta, order = -3)

makeobject.npnormcll(v, mu0, pi0, beta, order)

makeobject.npnormcvm(v, mu0, pi0, beta, order)

makeobject.npnormcvmw(v, mu0, pi0, beta, order = -3)

makeobject.npnormll(v, mu0, pi0, beta, order)

makeobject.npnormllw(v, mu0, pi0, beta, order = -3)

makeobject.nptll(v, mu0, pi0, beta, order)
}
\arguments{
\item{v}{the object either numeric or the implmented family}

\item{method}{An implemented family; see details}

\item{...}{other parameter passed to the constructor.}

\item{mu0}{A vector of support points}

\item{pi0}{A vector of weights corresponding to the support points}

\item{beta}{structual parameter.}

\item{order}{the parameter for the binned version.}
}
\description{
These functions are used to make the object for computing the non-paramtric mixing
distribution or estimating the proportion of zero using non-parametric methods.
}
\details{
This is a generic function for making the object for computing the non-parametric
mixing distribution or estimating the proportion of zero.

current implemented families are:
\itemize{
\item npnormll : normal density using maximum likelihood (Chapter 3). The default beta is 1.
\item npnormllw : Binned version of normal density using maximum likelihood (Chapter 6).
The default beta is 1 and the default order is -3.
\item npnormcvm : normal density using the Cramer-von Mises distance (Chapter 5). The default beta is 1.
\item npnormcvmw : Binned version of normal density using the Cramer-von Mises distance (Chapter 6).
The default beta is 1 and the default order is -3.
\item npnormad : normal density using the Anderson-Darling distance (Chapter 5). The default beta is 1.
\item npnormadw : Binned version of normal density using the Anderson-Darling distance (Chapter 6)
The default beta is 1 and the default order is -3.
\item nptll : t-density using maximum likelihood (Chapter 3). The default beta is infinity (normal distribution).
\item npnormcll : the one-parameter normal distribution used for approximating the sample
correlation coefficients using maximum likelihood. This does not have a
corresponding estimation of zero due to incompleted theory (Chapter 8).
There is no default beta. The structure beta is the number of observations.
}

The default method used is npnormll.

The detailed description of the npfixedcompRobject class is in \code{\link{npfixedcompRobject}}.
}
