% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcompR.R, R/npnormfccvm.R, R/npnormfcll.R
\name{makeobject}
\alias{makeobject}
\alias{makeobject.default}
\alias{makeobject.npnormcvm}
\alias{makeobject.npnormll}
\title{make object for computation}
\usage{
makeobject(v, mu0, pi0, beta)

\method{makeobject}{default}(v, mu0 = 0, pi0 = 0, beta = 1)

\method{makeobject}{npnormcvm}(v, mu0 = 0, pi0 = 0, beta = 1)

\method{makeobject}{npnormll}(v, mu0 = 0, pi0 = 0, beta = 1)
}
\arguments{
\item{v}{the object either numeric or the implmented family}

\item{mu0}{A vector of support points}

\item{pi0}{A vector of weights corresponding to the support points}

\item{beta}{structual parameter}
}
\description{
make object for computation
}
\details{
This is a S3 generic function for making the object for computing the non-parametric
mixing distribution

The default method used is npnormfcll.
}
