% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultility.R
\name{plotposteriormapping}
\alias{plotposteriormapping}
\title{plot the posterier map}
\usage{
plotposteriormapping(x, result, result2 = NULL, ...)
}
\arguments{
\item{x}{a vector of observations}

\item{result}{an object of class nspmix to show in the top}

\item{result2}{an object of class nspmix to show in the bottom. If NULL,
then the density in the bottom is not drawn.}

\item{...}{other parameter passed to \code{plot}}
}
\value{
none
}
\description{
plot the mapping between the original observations and its posterior mean
}
\details{
This function explicity considers that map between the transformed sample
correlation coefficients to the posterior mean of tanh(x) under normal cases.
\code{result} typically is the mixing distribution of the transformed sample
correlations and \code{result2} is the mixing distribution on the sample
correlation scale.
}
\examples{
n = 100; p = 50
X = matrix(rnorm(n * p), nrow = n, ncol = p)
r = covestEB(X)
x = makeobject(extractlower(r))
r1 = computemixdist(x)
plotposteriormapping(extractlower(r), r1)
}
